Move Downloader to nd4j-common; clean up old download code for MNIST
