+++ deeplearning4j/deeplearning4j-common/src/main/java/org/deeplearning4j/common/resources/DL4JResources.java
@@ -126,8 +126,4 @@ public class DL4JResources {
         f.mkdirs();
         return f;
     }
-
-
-
-
 }
