+++ deeplearning4j/deeplearning4j-data/deeplearning4j-datasets/src/main/java/org/deeplearning4j/base/IrisUtils.java
@@ -19,13 +19,12 @@
 package org.deeplearning4j.base;
 
 import org.apache.commons.io.IOUtils;
-import org.deeplearning4j.common.Downloader;
 import org.deeplearning4j.common.resources.DL4JResources;
 import org.deeplearning4j.common.resources.ResourceType;
 import org.nd4j.linalg.api.ndarray.INDArray;
 import org.nd4j.linalg.dataset.DataSet;
 import org.nd4j.linalg.factory.Nd4j;
-import org.nd4j.linalg.io.ClassPathResource;
+import org.nd4j.resources.Downloader;
 
 import java.io.File;
 import java.io.FileInputStream;
