+++ nd4j/nd4j-common/src/main/java/org/nd4j/resources/Downloader.java
@@ -1,4 +1,4 @@
-package org.deeplearning4j.common;
+package org.nd4j.resources;
 
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.codec.digest.DigestUtils;
